### 2024-12-11 #7585 @editorialbot
[PRE REVIEW]: RiverFlowDynamics v1.0: A Landlab component for computing two-dimensional river flow dynamics
**Submitting author:** @angelmons (<a href="http://orcid.org/0000-0002-7369-1602">Angel Monsalve</a>)
**Repository:** https://github.com/angelmons/landlab
**Branch with paper.md** (empty if default branch): riverFlowDynamics_dev
**Version:** v1.0
