### 2024-12-13 #7585 @angelmons
Hello @kthyng,
Thank you for your review and questions. I'd like to clarify that we are submitting a specific component called RiverFlowDynamics, which is integrated into the Landlab platform. You are correct that the code metrics shown above reflect the entire Landlab codebase rather than our specific component.

Regarding the submission requirements:

    1. Our component is covered under Landlab's MIT license, which is OSI-approved.
    2. The component is fully integrated into Landlab's official documentation system, including:
        - Detailed installation instructions as part of Landlab
        - Component-specific documentation
        - Comprehensive test suite
    3. Regarding the paper:

        - It is located at: https://github.com/landlab/landlab/blob/master/joss/in_preparation/river_flow_dynamics/paper.md
        - The current word count is approximately 1150 words, which we can adjust if needed during the review process
        - The paper and code demonstrate substantial scholarly effort, representing years of research and development
        - We clearly outline the research applications and potential impact

I noticed that the editorialbot is analyzing Pfeiffer et al. (2020) NetworkSedimentTransporter component in the published folder, rather than our RiverFlowDynamics component submission located in the in_preparation folder. This may explain the discrepancy in the code metrics being reported.

Would you like me to provide any additional information about our component or clarify any specific aspects of the submission?

Regarding potential reviewers: I would suggest: zsylvester, ebgoldstein, pfeiffea, nicgaspar
