### 2024-12-16 #7585 @angelmons
> @angelmons It's been awhile since I've seen a landlab submission come through, but am I correct in thinking that the number of lines of code shown above is for all of landlab, but you are submitting on a specific module? If so, please provide more information so we can find the correct numbers for the relevant work.

Hello @kthyng,
I hope you're having a good week. I wanted to gently follow up on our submission and check if you need any additional information from us regarding the RiverFlowDynamics component. We're happy to provide more details or clarification if needed.

Best regards.
