### 2024-12-18 #7585 @kthyng
@angelmons What I meant is where in the codebase can we find the code that is relevant to this submission so I can produce metrics for only the relevant part of the codebase?

I'm working on a solution for finding the correct paper.
