### 2024-12-18 #7585 @angelmons
Hello @kthyng
Thanks for the clarification. The RiverFlowDynamics component's code and related files are distributed across the following locations in the Landlab repository:

Main Component Source Code:
https://github.com/landlab/landlab/tree/master/src/landlab/components/river_flow_dynamics
Component Tests:
https://github.com/landlab/landlab/tree/master/tests/components/river_flow_dynamics
JOSS Paper:
https://github.com/landlab/landlab/tree/master/joss/in_preparation/river_flow_dynamics
Documentation & Tutorials:
https://github.com/landlab/landlab/tree/master/docs/source/tutorials/river_flow_dynamics

The primary implementation code is in the first link (source code). This should be the most relevant directory for code metrics calculations.
Please let me know if you need any additional information about specific files or if there's anything else I can help with to move the review process forward.
