### 2024-12-20 #7585 @kthyng
@angelmons Regarding the JOSS paper, can you either rename the old paper(s) (so they aren't paper.md) or make a temporary branch with only your paper in it. With either of those solutions, editorial bot should then be able to find your paper correctly.
