### 2025-01-03 #7585 @angelmons
@kthyng Thank you for the suggestion. I'm happy to inform you that the Landlab repository manager has already cleaned up the JOSS folder, and now our paper is the only one remaining in the "in_preparation" directory at:

https://github.com/landlab/landlab/tree/master/joss/in_preparation/river_flow_dynamics/paper.md

The file is still named paper.md as per the repository manager's recommendation. Since it's now the only paper in the in_preparation folder, would this solve the editorial bot's paper detection issue?
