### 2025-01-08 #7585 @kthyng
@angelmons I think you are missing a comma at the end of line 35: https://github.com/landlab/landlab/blob/master/joss/in_preparation/river_flow_dynamics/paper.bib
