### 2025-01-08 #7585 @angelmons
Hi @kthyng,

Thanks for catching the missing comma. To streamline the revision process, would you be willing to review the paper from my GitHub fork? It contains the exact same content as the Landlab master branch, with only the comma correction added. This would allow me to verify everything is in order before submitting an issue or pull request to the main Landlab repository.

The JOSS paper can be found here:
https://github.com/angelmons/landlab/blob/riverFlowDynamics_dev/joss/in_preparation/river_flow_dynamics

I just checked and the paper compiles all right. Please let me know
