### 2025-01-09 #7585 @angelmons
@kthyng Thanks! Yes, a branch will work. This is my branch:
https://github.com/angelmons/landlab/tree/riverFlowDynamics_dev
It is identical to the master branch, currently with the only exception being the missing comma.

As a summary:

Main Component Source Code:
https://github.com/angelmons/landlab/tree/riverFlowDynamics_dev/src/landlab/components/river_flow_dynamics
Component Tests:
https://github.com/angelmons/landlab/tree/riverFlowDynamics_dev/tests/components/river_flow_dynamics
JOSS Paper:
https://github.com/angelmons/landlab/tree/riverFlowDynamics_dev/joss/in_preparation/river_flow_dynamics
Documentation & Tutorials:
https://github.com/angelmons/landlab/tree/riverFlowDynamics_dev/docs/source/tutorials/river_flow_dynamics
