### 2025-01-21 #7585 @mcflugen
> Yes as long as you aren't conflicted!

I don't think I am. I'm one of the main developers of *Landlab*, so our project would benefit from this contribution but I'm not an author of this submission so it doesn't seem like a conflict.
