### 2025-01-21 #7585 @kthyng
Here is what I found running `cloc` on the component code specifically:

Since tests and docs are on top of this 1000 LOC, I think it passes substantial scholarly effort.
