### 2025-01-21 #7585 @kthyng
@angelmons paper comments:
* the references aren't working currently
* Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
* Please reduce to length of your paper to up to about 1000 words. Paper format information here: https://joss.readthedocs.io/en/latest/paper.html. You can check paper length (and a few other items) with the command `@editorialbot check repository` though note it needs to be the first item in a comment to run correctly. Currently it's 1220 words.
