### 2025-01-22 #7585 @angelmons
Hi @kthyng and @cheginit, I have addressed the issues you raised:

Fixed the capitalization in references using {} brackets
Reduced the paper length to under 1000 words
Corrected the citation formats

Let me know if there are any other comments or adjustments needed. I'll be happy to address them as soon as possible.
