### 2025-01-24 #7585 @cheginit
@angelmons Thanks for addressing the issues. I will be editing your submission. To begin, please provide a list of GitHub handles of potential reviewers without using `@` to prevent premature pinging. In addition to people in your own network, you can use JOSS's search [tool](https://reviewers.joss.theoj.org/reviewers).
