### 2025-01-30 #7585 @mcflugen
> Would you like to review this submission to the Journal for Open Source Software?

@cheginit Sure!
