### 2025-03-24 #7823 @angelmons
Thank you @ZhiLiHydro and @cheginit for your valuable input.

Before addressing the excellent suggestion raised by @ZhiLiHydro, I'd like to consolidate all comments and suggestions to handle them comprehensively. This approach will ensure that any overlapping suggestions are addressed efficiently.

@mcflugen and @xiaxilin, regarding the report in your last comment - could you clarify if this is a final version or still under development? I notice some unchecked items that are typically required when adding a component to the Landlab repository, so they are in fact already included in the component and I'd like to confirm the status before proceeding with any changes.

Once I have this information, I'll begin addressing all feedback systematically.
