### 2025-03-31 #7823 @mcflugen
@angelmons, @cheginit Review completed.

@angelmons Great work! The *RiverFlowDynamics* component addresses a critical gap in the *Landlab* component suite by providing a full solution of the depth-averaged *Saint-Venant equations*, enabling the simulation of a wide range of complex river and floodplain dynamics.

For your consideration, I have some general comments.

* There is a typo in the *Summary* section where *Landlab* is written as *LandLab*. I think that's the only occurrence but maybe worth double-checking.
* You mention that there are tutorials in the form of Jupyter notebooks that provide examples of how to use the *RiverFlowDynamics* component. It might be nice to describe how one might find them. That is, they are contained in the notebook gallery of the main *Landlab* documentation, and also as separate notebooks contained in the main *landlab* repository (they are not in a separate repository as stated in this article). No need to provide external links, just a general description of where they are.
* Consider elaborating on the development of future extensions to your component or how this component might be used with other *Landlab* components. For example, maybe coupling with a sediment transport component?
* Along the lines of the previous comment, maybe elaborate on why you chose to include this component as part of *Landlab* rather than writing it as a separate model. What is the advantage of having *RiverFlowDynamics* as a *Landlab* component?
* If possible, I think this article would benefit with a brief discussion of any validation or benchmarking that was done (e.g. comparison to analytic solutions). I believe this may be part of the unit tests, is that right?
* How does the *RiverFlowDynamics* component compare to comparable existing models (I'm not sure what those models are, but maybe TELEMAC or delft3D)?
* Because the model is being promoted for its efficiency, a few sentences with indicative timings or scaling behavior (even qualitative) would be useful.
* Consider a sentence or two about the component's limitations. For example, is this component able to work with all *Landlab* grid types (I don't believe so) or just uniform rectilinear grids? What about rectilinear grids with non-uniform spacing? What about curvilinear grids of quadrilaterals?
