### 2025-04-01 #7823 @xiaxilin
@angelmons

After reading the paper and the repository, I have a few general comments

1. I struggle to find the tutorial for the RiverDynamics component - it will be helpful to signpost where it is.
2. For statement of need, the paper mentions about Landlab - it is also useful to mention other models in addition to Landlab package for users to better understand the context.
