### 2025-04-02 #7823 @angelmons
# Responses to comments on the article

## **RiverFlowDynamics v1.0: A Landlab component for computing two-dimensional river flow dynamics**

### **Submitted to JOSS: Journal of Open Source Software**

By: A. Monsalve; S. Bernal; O. Link

We would like to express our sincere gratitude to the editor and reviewers for handling this revision and for their thoughtful comments and observations. Your feedback has significantly improved the quality and clarity of our paper while helping us better showcase the capabilities of **RiverFlowDynamics**.

We have addressed all the required observations while maintaining the restriction of 1000 words required by JOSS. More extended answers are provided in this detailed response document to fully explain our changes and reasoning, though the paper itself remains concise to comply with journal requirements.

Thank you again for your valuable contributions to this process.
