### 2025-04-02 #7823 @angelmons
## **Reviewer #1 – @ZhiLiHydro**

Thank you for thoroughly testing our installation process and working through the example notebooks. We appreciate your positive feedback about the structure and documentation. Your suggestion about including a more complex terrain example was excellent and has led to the development of a more comprehensive tutorial that better demonstrates the capabilities of **RiverFlowDynamics**.

1) *I went through the installation steps, successfully running the example jupyter notebooks. Some aspects, such as the installation instructions and community guidelines, are great—largely because landlab as a whole is already well-structured.*

R: Thank you for your positive feedback on our installation instructions and community guidelines. We appreciate your thorough review and confirmation that the examples run successfully.

2) *My only minor suggestion is to include a more complex terrain. The current dataset, DEM-kootenai_37x50_1x1.asc, is relatively simple for testing. A more complex terrain would allow for better evaluation, and if included, additional 2D visualizations in the tutorial notebook would be beneficial.*

R: Thank you for this excellent suggestion. We agree that including a more complex terrain would better showcase the capabilities of **RiverFlowDynamics**. We have developed a new comprehensive tutorial with a more complex topography and enhanced 2D visualizations, which is now available in the **Landlab** notebook gallery.
