### 2025-04-02 #7823 @angelmons
## **Reviewer #2 – @mcflugen**

Thank you for your detailed and constructive feedback that helped us significantly improve the paper. Your suggestions regarding clarification of tutorial locations, discussion of future extensions, advantages of Landlab integration, validation procedures, model comparisons, and limitations have all contributed to a more complete presentation of **RiverFlowDynamics**. We especially appreciate your recognition of how this component addresses a critical gap in the Landlab ecosystem.

*Great work! The RiverFlowDynamics component addresses a critical gap in the Landlab component suite by providing a full solution of the depth-averaged Saint-Venant equations, enabling the simulation of a wide range of complex river and floodplain dynamics.*

*For your consideration, I have some general comments.*

*1) There is a typo in the Summary section where Landlab is written as LandLab. I think that's the only occurrence but maybe worth double-checking.*

R: Thank you for catching this error. We have corrected the typo throughout the document to consistently use "Landlab" rather than "LandLab".

*2) You mention that there are tutorials in the form of Jupyter notebooks that provide examples of how to use the RiverFlowDynamics component. It might be nice to describe how one might find them. That is, they are contained in the notebook gallery of the main Landlab documentation, and also as separate notebooks contained in the main landlab repository (they are not in a separate repository as stated in this article). No need to provide external links, just a general description of where they are.*

R: Thank you for this clarification. You are correct that the tutorials are part of the Landlab notebook gallery within the main repository, not in a separate repository as stated. We have updated this information in the paper.

*3) Consider elaborating on the development of future extensions to your component or how this component might be used with other Landlab components. For example, maybe coupling with a sediment transport component?*

R: Thank you for suggesting we elaborate on future extensions. We have added information on potential coupling with sediment transport components. We added to Statement of need: "The model's integration into Landlab's component framework enables future coupling with sediment transport components to simulate morphodynamic processes and assess impacts on aquatic habitat and riverine vegetation dynamics under changing flow conditions."

*4) Along the lines of the previous comment, maybe elaborate on why you chose to include this component as part of Landlab rather than writing it as a separate model. What is the advantage of having RiverFlowDynamics as a Landlab component?*

R: Thank you for this insightful comment. We have added text explaining the advantages of developing **RiverFlowDynamics** as a Landlab component rather than a standalone model. We added to Statement of need: "Integration with Landlab leverages its established grid structure, visualization tools, and interoperability with other components, facilitating multi-process simulations and reducing development overhead."

*5) If possible, I think this article would benefit with a brief discussion of any validation or benchmarking that was done (e.g. comparison to analytic solutions). I believe this may be part of the unit tests, is that right?*

R: Thank you for highlighting the importance of validation. We have included a brief mention of the validation against analytical solutions that is part of our testing framework. We added to the Summary: "The model has been validated through a comprehensive test suite including analytical solutions, numerical stability tests, sensitivity analyses, mass conservation checks, and boundary condition validations."

*6) How does the RiverFlowDynamics component compare to comparable existing models (I'm not sure what those models are, but maybe TELEMAC or delft3D)?*

R: Thank you for suggesting comparisons with existing models. We have added a brief comparison with TELEMAC and Delft3D. We added to Statement of need: "Unlike commercial alternatives such as TELEMAC and Delft3D, **RiverFlowDynamics** provides an open-source, Python-based solution that is more accessible to researchers and educators while maintaining comparable numerical accuracy."

*7) Because the model is being promoted for its efficiency, a few sentences with indicative timings or scaling behavior (even qualitative) would be useful.*

R: Thank you for this suggestion. While we mention computational efficiency in the context of our numerical schemes (semi-implicit time stepping, conjugate gradient solver), performance benchmarking was not a primary focus of this initial release. The efficiency advantages mentioned in the paper relate primarily to the numerical properties (larger allowable time steps, rapid convergence) rather than specific computational performance claims. We plan to include more detailed performance analysis in future work, but have chosen to maintain focus on the model's capabilities and integration with Landlab in this paper given the word limit constraints.

*8) Consider a sentence or two about the component's limitations. For example, is this component able to work with all Landlab grid types (I don't believe so) or just uniform rectilinear grids? What about rectilinear grids with non-uniform spacing? What about curvilinear grids of quadrilaterals?*

R: Thank you for suggesting we address limitations. We have added clarification about grid compatibility to the Summary: "The current implementation supports uniform rectilinear grids only; planned extensions include support for non-uniform rectilinear and curvilinear meshes."
