### 2025-04-02 #7823 @angelmons
## **Reviewer #3 – xiaxilin**

Thank you for your insightful comments regarding the accessibility of tutorials and the need for broader context beyond the Landlab ecosystem. These suggestions have helped us improve both the documentation and the positioning of our work within the field of hydraulic modeling software. We have clarified the tutorial locations and expanded our discussion to include comparisons with other established models.

*After reading the paper and the repository, I have a few general comments*

*1) I struggle to find the tutorial for the RiverDynamics component - it will be helpful to signpost where it is.*

R: Thank you for noting the difficulty in finding the tutorial. We have clarified the location of tutorials in our paper as suggested in our response to Reviewer #2's second comment. The tutorials are part of the Landlab notebook gallery within the main repository, specifically located at `docs/source/tutorials/river_flow_dynamics`. They can be accessed directly through the GitHub repository at:
- https://github.com/landlab/landlab/tree/master/docs/source/tutorials/river_flow_dynamics
- https://github.com/angelmons/landlab/tree/riverFlowDynamics_dev/docs/source/tutorials/river_flow_dynamics

*2) For statement of need, the paper mentions about Landlab - it is also useful to mention other models in addition to Landlab package for users to better understand the context.*

R: Thank you for this valuable suggestion to provide broader context. We have expanded our Statement of need to include comparisons with models outside the Landlab ecosystem. Specifically, we've added a comparison with commercial hydraulic modeling software like TELEMAC and Delft3D, highlighting how **RiverFlowDynamics** offers a more accessible open-source alternative while maintaining comparable numerical accuracy through similar finite volume schemes and semi-implicit time stepping. As suggested in our response to Reviewer #2's Comment 6, we've added the following text to the Statement of need section: "Unlike commercial alternatives such as TELEMAC and Delft3D, **RiverFlowDynamics** provides an open-source, Python-based solution that is more accessible to researchers and educators while maintaining comparable numerical accuracy." This comparison helps position our tool in the broader landscape of hydraulic modeling software beyond just the Landlab framework.
