### 2025-04-03 #7823 @angelmons
@cheginit, @mcflugen, @ZhiLiHydro, @xiaxilin

On my side now, I will update the paper and ensure the new tutorial is fully available and thoroughly tested within the repository. Once verified for functionality and clarity, I will push these changes to the GitHub repository. This will make the improvements accessible to all users and reviewers, completing the revision process as requested.

Regarding the paper revisions, please note that we've added only the most critical information requested, as JOSS requires papers to be less than 1000 words. More comprehensive descriptions would exceed this limit by a significant margin. We've carefully balanced addressing your valuable feedback while maintaining compliance with the journal's concise format requirements.
