### 2025-04-05 #7823 @cheginit
@mcflugen Thanks for providing feedback. However, please follow the instructions in the first comment of this issue (or my [comment](https://github.com/openjournals/joss-reviews/issues/7823#issuecomment-2671567629)) to create a checklist, then mark each item if you think the submission satisfies it. This is a required step for JOSS submissions to be accepted. Thanks for your help with this.
