### 2025-04-05 #7823 @mcflugen
@cheginit I thought I had [already done that](https://github.com/openjournals/joss-reviews/issues/7823#issuecomment-2672221859). Is there something more I need to do? After I go over the latest changes, I'll check the remaining boxes.
