### 2025-04-24 #7823 @angelmons
Hi @cheginit,

Thanks for following up. I'm making good progress and anticipate completing the review by Friday. I've already tested and checked a dam break case, which is very demanding in terms of numerical implementation and it is working correctly, and I'm currently verifying one remaining test cases to ensure everything is functioning as expected.

I'll provide a full update once all checks are complete. Let me know if you have any specific concerns in the meantime. The paper itself is complete as it was discussed above.

Best regards, Angel
