### 2025-04-28 ca38717
Update JOSS paper and tutorials based on reviewer feedback
- Fix typo: "LandLab" -> "Landlab" throughout the document
- Clarify tutorial location in Landlab documentation
- Add validation information to Summary section
- Add grid limitations and future extensions information
- Add comparison to existing models (TELEMAC, Delft3D)
- Add details on Landlab integration advantages
- Expand Statement of need with impact on aquatic habitat and vegetation

- Expand tutorials from 2 to 5, including:
  - Basic channel flow tutorial
  - Complex terrain flow with Kootenai River DEM
  - In-depth river flow
[…]
