+++ joss/in_preparation/river_flow_dynamics/paper.bib
@@ -1,13 +1,13 @@
 @article{hobley:2017,
   title={
-    Creative computing with Landlab: an open-source toolkit for building,
+    {Creative} computing with {Landlab}: an open-source toolkit for building,
     coupling, and exploring two-dimensional numerical models of
-    Earth-surface dynamics
+    {Earth}-surface dynamics
   },
   author={
-    Hobley, Daniel EJ and Adams, Jordan M and Nudurupati, Sai Siddhartha and
-    Hutton, Eric WH and Gasparini, Nicole M and Istanbulluoglu, Erkan and
-    Tucker, Gregory E
+    Hobley, Daniel E. J. and Adams, Jordan M. and Nudurupati, Sai Siddhartha and
+    Hutton, Eric W. H. and Gasparini, Nicole M. and Istanbulluoglu, Erkan and
+    Tucker, Gregory E.
   },
   journal={Earth Surface Dynamics},
   volume={5},
@@ -20,39 +20,33 @@
 }
 
 @article{barnhart:2020,
-  title={Landlab v2. 0: a software package for Earth surface dynamics},
+  title={{Landlab} v2.0: a software package for {Earth} surface dynamics},
   author={
-    Barnhart, Katherine R and Hutton, Eric WH and Tucker, Gregory E and
-    Gasparini, Nicole M and Istanbulluoglu, Erkan and Hobley, Daniel EJ and
-    Lyons, Nathan J and Mouchene, Margaux and Nudurupati, Sai Siddhartha and
-    Adams, Jordan M and others
+    Barnhart, Katherine R. and Hutton, Eric W. H. and Tucker, Gregory E. and
+    Gasparini, Nicole M. and Istanbulluoglu, Erkan and Hobley, Daniel E. J. and
+    Lyons, Nathan J. and Mouchene, Margaux and Nudurupati, Sai Siddhartha and
+    Adams, Jordan M. and others
   },
   journal={Earth Surface Dynamics},
   volume={8},
   number={2},
   pages={379--397},
   year={2020},
-  publisher={Copernicus GmbH}
-  url = {https://esurf.copernicus.org/articles/8/379/2020/},
-  doi = {10.5194/esurf-8-379-2020}
+  publisher={Copernicus GmbH},
+  url={https://esurf.copernicus.org/articles/8/379/2020/},
+  doi={10.5194/esurf-8-379-2020}
 }
 
 @article{casulli_semi-implicit_1999,
-	title = {A semi-implicit finite difference method for non-hydrostatic, free-surface flows},
-	volume = {30},
-	copyright = {http://doi.wiley.com/10.1002/tdm\_license\_1.1},
-	issn = {0271-2091, 1097-0363},
-	url = {https://onlinelibrary.wiley.com/doi/10.1002/(SICI)1097-0363(19990630)30:4<425::AID-FLD847>3.0.CO;2-D},
-	doi = {10.1002/(SICI)1097-0363(19990630)30:4<425::AID-FLD847>3.0.CO;2-D},
-	abstract = {In this paper a semi-implicit finite difference model for non-hydrostatic, free-surface flows is analyzed and discussed. It is shown that the present algorithm is generally more accurate than recently developed models for quasi-hydrostatic flows. The governing equations are the free-surface Navier – Stokes equations defined on a general, irregular domain of arbitrary scale. The momentum equations, the incompressibility condition and the equation for the free-surface are integrated by a semi-implicit algorithm in such a fashion that the resulting numerical solution is mass conservative and unconditionally stable with respect to the gravity wave speed, wind stress, vertical viscosity and bottom friction. Copyright © 1999 John Wiley \& Sons, Ltd.},
-	language = {en},
[…]
