### 2025-04-28 8d95888
Update paper.bib
Updates some missing doi
