+++ joss/in_preparation/river_flow_dynamics/paper.bib
@@ -58,7 +58,7 @@
   pages={56--74},
   year={1990},
   publisher={Elsevier},
-  doi={10.1016/0021-9991(90)90091-5}
+  doi={10.1016/0021-9991(90)90091-E}
 }
 
 @article{robert1985semi,
@@ -164,7 +164,7 @@
   doi={10.5194/gmd-10-1645-2017}
 }
 
-@article{de_Almeida2012,
+@article{de2012improving,
   title={Improving the stability of a simple formulation of the shallow water equations for 2-{D} flood modeling},
   author={de Almeida, Gustavo A. M. and Bates, Paul and Freer, Jim E. and Souvignet, Maxime},
   journal={Water Resources Research},
@@ -179,6 +179,7 @@
   author={Andersson, Bengt and Andersson, Ronnie and Håkansson, Love and Mortensen, Mikael and Sudiyo, Rahman and Van Wachem, Berend},
   year={2011},
   publisher={Cambridge University Press},
+  doi={10.1017/CBO9781139093590},
   isbn={978-0521189613}
 }
 
@@ -187,6 +188,7 @@
   author={Fletcher, Clive A. J.},
   year={2012},
   publisher={Springer Science \& Business Media},
+  doi={10.1007/978-3-642-58239-4},
   isbn={978-3642584947}
 }
 
