### 2025-04-28 #7823 @angelmons
Hello @cheginit, @mcflugen, @ZhiLiHydro, and @xiaxilin,

I'm pleased to inform you that the updated version of our paper is now available, addressing all the points raised during the review process. The specific issues addressed in the paper are explained in detail in our comments above.

Regarding the tutorials, we've completely redesigned them to enhance the content and explanations. What were initially 2 tutorials have now been expanded into 5 comprehensive tutorials, with special attention to @ZhiLiHydro's suggestion about including more complex terrain examples.

The 5 tutorials now available include:

i) **Basic Channel Flow** - A foundational tutorial demonstrating water flow in a simple rectangular channel, introducing the core concepts and functionality of RiverFlowDynamics.

ii) **Complex Terrain Flow** - An enhanced tutorial using the Kootenai River DEM with improved visualizations showing how water flows over natural topography.

iii) **Mastering River Flow Dynamics** - A comprehensive tutorial that bridges theory and simulation, providing deeper insights into the hydraulic principles with extensive visualization and analysis.

iv) **Circular Dam Break** - A new tutorial demonstrating the component's ability to simulate complex hydraulic phenomena like a dam break over a flat surface.

v) **Green River Case Study** - A completely new tutorial using high-resolution DEM data of the Green River in Utah, showcasing the component's capabilities with real-world complex terrain, including extensive cross-sectional analysis.

We've significantly improved the visualizations within the Jupyter Notebooks, though as noted, for truly high-resolution or specialized visualizations, exporting to GIS would be recommended.

Thank you again for your valuable feedback which has helped us greatly improve both the paper and the accompanying tutorials.

Please let me know what the next step is.
