### 2025-04-28 #7823 @xiaxilin
@angelmons Thank you for making these changes. I tried to run the tutorial on Google Colab. The installation was fine by using 'pip install landlab', but I encountered an error when running the first block of the code. The error message was

ImportError                               Traceback (most recent call last)
[<ipython-input-2-41f891bebb5e>](https://jkyy6mjysvi-496ff2e9c6d22116-0-colab.googleusercontent.com/outputframe.html?vrz=colab_20250425-101002_RC00_751415424#) in <cell line: 0>()
      5
      6 from landlab import RasterModelGrid
----> 7 from landlab.components import RiverFlowDynamics
      8 from landlab.io import esri_ascii

ImportError: cannot import name 'RiverFlowDynamics' from 'landlab.components' (/usr/local/lib/python3.11/dist-packages/landlab/components/__init__.py)

Did I miss anything?

Please also note that both Telemac and Delft3D are open-source software, so this needs to be corrected in the paper.
