### 2025-04-28 #7823 @angelmons
Hi @xiaxilin,

Thank you for trying out the tutorial and providing this feedback! Regarding the RiverFlowDynamics import error - this could be due to a version mismatch. The component might be named differently in your installed version of Landlab or require an additional installation step. My guess is that there is an error in the python environment.

Could you check which version of Landlab you have installed with pip show landlab? Also, have you tried running any other Landlab tutorials successfully? This information would help us troubleshoot the specific issue.

I just performed a fresh installation from Landlab master last night and was able to run all tutorials without any problems. Both @ZhiLiHydro and @mcflugen have also reported successfully running the tutorials, which suggests this might be related to your specific installation or environment.

Regarding Telemac and Delft3D - thank you for pointing out that they are open-source software. I'll update that in the paper today.

@cheginit - What would you advise for addressing new comments? Should I wait to accumulate a comprehensive list or address them one by one as they appear? How is the typical review process at this 2nd iteration?
