### 2025-04-29 #7823 @xiaxilin
Hi @angelmons, the version I installed was landlab 2.9.2. I tried to do it locally by creating a new conda environment, and it was still not working. I just used 'pip install landlab'. Did I have to install from source? Could you share your steps of installation?
