### 2025-04-29 #7823 @angelmons
Hi @cheginit, sounds good. No major issues have been reported. The new tutorials were already checked by @ZhiLiHydro, so I assume they are working well not only on my side.

Hi @cheginit, yes, I can help you with the installation. Most of the time the instructions on the Landlab website will work without any problems, but I have experienced small issues due to Python updates or Microsoft Visual dependencies. I would like to start a fresh installation to help you with clear steps. What OS do you use? Linux, Windows, or Mac?
