### 2025-04-29 #7823 @xiaxilin
@angelmons Thank you, if you could investigate the installation on Google Colab that would be really useful. I used the default setting using the free version of Google Colab. It would be good to resolve the issue on the platform because may people would use it.
