### 2025-04-29 #7823 @xiaxilin
@angelmons Thank you for your clarification. I did test it locally on Mac. I created a new conda environment with Python 3.10, then used 'pip install landlab' to install by following the instructions. But I still faced the same error

from landlab.components import RiverFlowDynamics
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
ImportError: cannot import name 'RiverFlowDynamics' from 'landlab.components' (/Users/xiaxz/opt/anaconda3/envs/landlab/lib/python3.10/site-packages/landlab/components/__init__.py
