### 2025-04-29 #7823 @ZhiLiHydro
> [@angelmons](https://github.com/angelmons) Thank you for your clarification. I did test it locally on Mac. I created a new conda environment with Python 3.10, then used 'pip install landlab' to install by following the instructions. But I still faced the same error
>
> from landlab.components import RiverFlowDynamics Traceback (most recent call last): File "", line 1, in ImportError: cannot import name 'RiverFlowDynamics' from 'landlab.components' (/Users/xiaxz/opt/anaconda3/envs/landlab/lib/python3.10/site-packages/landlab/components/**init**.py
>
> After checking __init__.py, I found that the RiverFlowDynamics isn't there. So it seems that the river flow component isn't in the official PyPI repository. Therefore I would recommend providing additional installation documentation specific for this river flow component.

(just in case this is helpful) I installed the`riverFlowDynamics_dev` branch from `angelmons`'s fork:
`git clone -b riverFlowDynamics_dev https://github.com/angelmons/landlab.git`
`python -m pip install -e .`
Next, I ran the notebooks in https://github.com/angelmons/landlab/tree/riverFlowDynamics_dev/docs/source/tutorials/river_flow_dynamics
