### 2025-04-29 #7823 @angelmons
Thanks @ZhiLiHydro, to complement:

Hi @xiaxilin,

Thank you for providing those additional details about your testing process. I understand your frustration with the import error, and I appreciate you taking the time to investigate further.

The issue you're encountering is because RiverFlowDynamics is currently only available in the developer version of Landlab, not in the standard PyPI release. This is actually expected behavior, as new components typically go through Landlab's development process before being included in the official release.

For accessing this component, you'll need to install the developer version as mentioned in the Landlab documentation:
https://landlab.readthedocs.io/en/v2.9.2/install/developer_install.html#install

I've just completed a fresh installation on my Mac following these developer installation instructions, and was able to import RiverFlowDynamics without issues. The installation process may prompt you to install additional dependencies, but it should handle these automatically.

To run the tutorials specific to this component, please use the ones in this review repository:
**Repository:** https://github.com/angelmons/landlab
**Branch:** riverFlowDynamics_dev

Once this paper is accepted, I plan to submit a pull request to integrate this component into the official Landlab repository, which would make it available through standard installation methods in future releases.

Please let me know if you encounter any further issues with the developer installation, and I'd be happy to help troubleshoot.
