### 2025-05-08 #7823 @angelmons
Hi @cheginit and @xiaxilin

I hope this message finds you well. I wanted to follow up regarding my paper submission, which has been under review since February 20. It's been over a week since my last response to reviewer comments, and I haven't received any further updates on the status of the review process.

I'm pleased to note that the paper has already received two positive reviews, which seems to indicate we may be approaching the final stages of the review process. As this is a short paper presenting a specific Landlab component, I'm hoping we might be able to move toward a resolution soon.

I understand review processes take time and appreciate the thorough evaluation being conducted. I want to emphasize that I'm fully committed to helping at any step that I can to move this forward quickly. If there are any remaining concerns or questions, I'm ready to address them promptly.

Could you please provide an update on the current status of the review and any estimated timeline for the next steps?

Thank you for your continued consideration and for the valuable feedback provided thus far.

Best regards,

Angel
