### 2025-05-08 #7823 @xiaxilin
@angelmons Thank you for providing further information about the installation. I have now managed to run it on Windows with WSL but the installation still failed on Mac OS.

I have run through all the tutorials. They all run without problems. However, for tutorial 4 about the circular dam break, the results are asymmetrical with oscillations on the north and east sides of the shock wave fronts. The results should be fully symmetrical in this case. The results seem to suggest that there is an issue with the implementation of the numerical method. So please check this. An example of the same case study can be found in the given reference, which gives fully symmetrical water depth profile.

Reference:
Wang, Y., Liang, Q., Kesserwani, G. and Hall, J.W., 2011. A 2D shallow flow model for practical dam-break simulations. Journal of Hydraulic Research, 49(3), [pp.307-316.]
