### 2025-05-23 #7823 @angelmons
Hello @cheginit and @xiaxilin,

Thank you for bringing this important observation to our attention and for providing the reference for comparison. I appreciate your thorough testing of the tutorials and the detailed feedback regarding the asymmetrical results in tutorial 4.

We have been investigating this specific behavior in the circular dam break case. Our preliminary analysis suggests that the observed asymmetry and oscillations may be related to numerical precision and the accumulation of floating-point rounding errors over the 71 time steps, rather than a fundamental implementation error in the numerical method itself. However, we will keep testing some ideas to ensure the robustness of the implementation.

We are currently conducting additional tests to:
1. Verify the numerical scheme implementation against the reference solution
2. Assess the impact of different numerical precision settings
3. Evaluate potential mitigation strategies for reducing numerical artifacts

We will provide a comprehensive update once we have completed our analysis and either confirmed that the observed behavior is within acceptable numerical limits or identified and addressed any implementation issues.

Thank you for your patience as we work through this important validation step.

Best regards, AM
