### 2025-06-05 fc96abf
Fix asymmetry issues in shallow water flow solver and enhance Tutorial 4
- Fix processing order bias in water depth calculation causing asymmetrical results
- Replace sequential core node processing with simultaneous vectorized computation
- Improve numerical stability in path line tracing with robust safe thresholds
- Enhance symmetric boundary and corner node treatment
- Add dynamic visualization capabilities to Tutorial 4 with animation output
- Achieve 1000x improvement in symmetry preservation (asymmetry reduced from 0.098 to 10^-4)
- Maintain scientific-grade CFD symmetry standards for c
[…]
