+++ src/landlab/components/river_flow_dynamics/river_flow_dynamics.py
@@ -723,17 +723,43 @@ class RiverFlowDynamics(Component):
                 self._v_vel_of_particle < 1e-10, 0, self._v_vel_of_particle
             )
 
-            ### Calculation accoss x-direction
-            # Avoiding divisions by zero
-            tempCalc1 = np.where(
-                self._u_vel_of_particle == 0, 9999, self._u_vel_of_particle
+            # Small velocity handling for both components
+            self._u_vel_of_particle = np.where(
+                np.abs(self._u_vel_of_particle) < 1e-10, 
+                0, self._u_vel_of_particle
             )
-            tempCalc2 = np.where(u_vel_at_x1 == 0, 9999, u_vel_at_x1)
-            tempCalc3 = np.where(gradient_x_direction == 0, 9999, gradient_x_direction)
-            TAUx = (1 / tempCalc3) * np.log(abs(tempCalc1 / tempCalc2))
+            self._v_vel_of_particle = np.where(
+                np.abs(self._v_vel_of_particle) < 1e-10, 
+                0, self._v_vel_of_particle
+            )
+
+            ### Calculation across x-direction
+            safe_threshold = 1e-12
+            tempCalc1 = np.where(
+                np.abs(self._u_vel_of_particle) < safe_threshold, 
+                np.sign(self._u_vel_of_particle) * safe_threshold, 
+                self._u_vel_of_particle
+            )
+            tempCalc2 = np.where(
+                np.abs(u_vel_at_x1) < safe_threshold, 
+                np.sign(u_vel_at_x1) * safe_threshold, 
+                u_vel_at_x1
+            )
+            tempCalc3 = np.where(
+                np.abs(gradient_x_direction) < safe_threshold, 
+                np.sign(gradient_x_direction) * safe_threshold, 
+                gradient_x_direction
+            )
+
+            with np.errstate(divide='ignore', invalid='ignore'):
+                ratio = tempCalc1 / tempCalc2
+                ratio = np.where(np.isfinite(ratio), ratio, 1.0)
+                ratio = np.clip(ratio, safe_threshold, 1/safe_threshold)
+                TAUx = (1 / tempCalc3) * np.log(np.abs(ratio))
 
             # Calculation when gradient is equal to zero
-            tempCalc4 = abs((self._x_of_particle - x_at_x1) / tempCalc2)
+            with np.errstate(divide='ignore', invalid='ignore'):
+                tempCalc4 = abs((self._x_of_particle - x_at_x1) / tempCalc2)
             TAUx = np.where(gradient_x_direction == 0, tempCalc4, TAUx)
 
             # Calculation when:
@@ -741,25 +767,43 @@ class RiverFlowDynamics(Component):
             # b) Uxp,Vyp = 0,
             # c) Ux1,Vy1 = 0, and
             # d) Uxp/Ux1, Vxp/Vy1 = -1
-            tempCalc5 = self._u_vel_of_particle / tempCalc2
+            with np.errstate(divide='ignore', invalid='ignore'):
+                tempCalc5 = self._u_vel_of_particle / tempCalc2
             TAUx = np.where(tempCalc5 == 1, tempCalc4, TAUx)
             TAUx = np.where(self._u_vel_of_particle == 0, remaining_time, TAUx)
             TAUx = np.where(u_vel_at_x1 == 0, remaining_time, TAUx)
             TAUx = np.where(tempCalc5 < 0, remaining_time, TAUx)
-            TAUx = np.where(TAUx > self._dt, self._dt, TAUx)
-            TAUx = np.where(TAUx < 0, 0, TAUx)
+            TAUx = np.where(TAUx >= self._dt, self._dt, TAUx)
[…]
