+++ docs/source/tutorials/river_flow_dynamics/river_flow_dynamics_tutorial4_plot.py
@@ -1,0 +1,414 @@
+# -*- coding: utf-8 -*-
+"""
+Visualization for landlab RiverFlowDynamics simulations
+
+@author: angel
+"""
+import matplotlib.pyplot as plt
+import numpy as np
+from mpl_toolkits.mplot3d import Axes3D
+from matplotlib import cm
+import matplotlib.colors as colors
+from PIL import Image
+import glob
+import os
+
+# Additional imports for Jupyter notebook display
+try:
+    from IPython.display import Image as IPImage, display
+    IPYTHON_AVAILABLE = True
+except ImportError:
+    IPYTHON_AVAILABLE = False
+    print("Warning: IPython not available. GIF display in notebook will not work.")
+    
+# Import required libraries
+try:
+    from scipy.interpolate import griddata
+    from scipy.ndimage import gaussian_filter
+    SCIPY_AVAILABLE = True
+except ImportError:
+    SCIPY_AVAILABLE = False
+    print("Warning: scipy not available. Basic plotting will be used.")
+
+
+def get_nodes(grid, nrows, ncols):
+    """
+    Get nodes along the cross section.
+    
+    Parameters:
+    -----------
+    grid : RasterModelGrid
+        Landlab grid object
+    nrows, ncols : int
+        Grid dimensions
+        
+    Returns:
+    --------
+    plot_nodes : array
+        Node indices along water
+    plot_distances : array
+        Distances along water
+    """
+    plot_nodes = []
+    plot_distances = []
+    
+    # Calculate plot endpoints
+    x_start, y_start = grid.x_of_node[ncols - 1], grid.y_of_node[ncols - 1]  # Upper-left
+    x_end, y_end = grid.x_of_node[nrows * ncols - ncols], grid.y_of_node[nrows * ncols - ncols]  # Lower-right
+    
+    # Total plot distance
+    total_distance = np.sqrt((x_end - x_start)**2 + (y_end - y_start)**2)
+    
+    # Sample points along plot
+    n_samples = min(nrows, ncols)
+    
+    for i in range(n_samples):
+        t = i / (n_samples - 1)  # Parameter from 0 to 1
+        x_target = x_start + t * (x_end - x_start)
+        y_target = y_start + t * (y_end - y_start)
+        
+        # Find closest node
+        distances = np.sqrt((grid.x_of_node - x_target)**2 + (grid.y_of_node - y_target)**2)
+        closest_node = np.argmin(distances)
+        
+        plot_nodes.append(closest_node)
+        plot_distances.append(t * total_distance)
+    
+    return np.array(plot_nodes), np.array(plot_distances)
+
+
+def create_combined_plot(grid, timestep, field_name="surface_water__depth", 
+                        figsize=(18, 7), colormap='viridis'):
+    """
+    Create combined 3D and cross-section plot for RiverFlowDynamics visualization.
+    
+    Parameters:
+    -----------
+    grid : RasterModelGrid
+        Landlab grid object
+    timestep : int
+        Current timestep number
+    field_name : str
+        Field to plot (default: "surface_water__depth")
+    figsize : tuple
+        Figure size (width, height) - increased width for better spacing
+    colormap : str
+        Colormap name
+    """
+    
+    if not SCIPY_AVAILABLE:
+        print("Warning: scipy not available. Using basic 3D rendering.")
+    
+    # Get data
+    field_data = grid.at_node[field_name]
+    nrows, ncols = grid.shape
+    
+    # Get plot cross-section data
+    plot_nodes, plot_distances = get_nodes(grid, nrows, ncols)
+    
+    # Create figure with larger width and strategic layout
+    fig = plt.figure(figsize=figsize)
+    
[…]
