### 2025-06-21 #7823 @angelmons
Dear @mcflugen,
I hope this message finds you well. I wanted to follow up regarding the remaining unchecked items in the review checklist and discuss how we might move forward with the publication.
The outstanding items are:

i) Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
ii) Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
iii) Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
iv) State of the field: Do the authors describe how this software compares to other commonly-used packages?

Based on the recent article revisions, the software's implementation within the Landlab framework, and the positive reviews from the other two reviewers (who have already checked off these items), I believe these aspects have been adequately addressed.
However, I'm completely open to making any additional improvements you might suggest. Could you please let me know if there are specific areas within these four categories that you'd like me to elaborate on or modify? I'm committed to ensuring the work meets all publication standards.
Thank you for your time and thoughtful review process.
Best regards,
AM
