### 2025-06-21 #7823 @cheginit
@angelmons the paper reads well and has all the required information except for a description of high-level functionalities of the software as documented in JOSS's [website](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain):

> A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.

The paper currently contains too many technical details on the methodology but does not explain the functionalities of the software. Please reduce the technical details and add a brief demonstration section to discuss how the software can be used to run an example model and show a sample plot. For example, you can check out [this](https://joss.theoj.org/papers/10.21105/joss.08061) published paper.
