### 2025-06-23 6960cbf
Update paper.md
Updating paper - Getting it to 1000 words.
