+++ joss/in_preparation/river_flow_dynamics/paper.md
@@ -36,7 +36,7 @@ bibliography: paper.bib
 ---
 # Summary
 
-Numerical modeling of surface water flow is a critical tool in hydrology, hydraulics, and environmental science. These models play a crucial role in predicting and analyzing flow patterns in rivers, flood plains, and coastal areas, informing decisions in water resource management, flood risk assessment, and ecosystem conservation. This paper introduces a novel two-dimensional flow model, RiverFlowDynamics, developed as a component of the Landlab Python Package [@hobley:2017;@barnhart:2020;@Hutton:2020;@Hutton:2020], designed to simulate the behavior of rivers and streams under various flow conditions over natural and artificial topography.
+RiverFlowDynamics enables researchers to simulate how water flows through rivers, streams, and flood plains using realistic physics. The software can predict water depths, flow velocities, and how water dynamics evolve over complex terrain, making it valuable for flood risk assessment, environmental studies, and water resource management. Numerical modeling of surface water flow is a critical tool in hydrology, hydraulics, and environmental science. These models play a crucial role in predicting and analyzing flow patterns in rivers, flood plains, and coastal areas, informing decisions in water resource management, flood risk assessment, and ecosystem conservation. This paper introduces a novel two-dimensional flow model, RiverFlowDynamics, developed as a component of the Landlab Python Package [@hobley:2017;@barnhart:2020;@Hutton:2020;@Hutton:2020], designed to simulate the behavior of rivers and streams under various flow conditions over natural and artificial topography.
 
 RiverFlowDynamics is founded on the depth-averaged Saint-Venant equations, also known as the shallow water equations [@casulli1990semi;@casulli_semi-implicit_1999]. These equations, derived from the Navier-Stokes equations for incompressible flow, are simplified by integrating over the water depth. This approach assumes that vertical accelerations are negligible compared to horizontal ones, a reasonable approximation for many surface water flows. The governing equations consist of continuity and momentum balance equations in two dimensions, capturing the essential dynamics of free-surface flows.
 
@@ -62,9 +62,46 @@ Source code for RiverFlowDynamics is available as part of the Landlab Python pac
 
 RiverFlowDynamics is a Python-based 2D flow model developed as a component of the Landlab framework, addressing a critical gap in the modeling of complex river systems and flood dynamics. Prior to RiverFlowDynamics, Landlab lacked a comprehensive 2D flow model capable of handling fully advective-dominated problems, particularly in rivers with complex topographies. This limitation hindered accurate simulations of diverse flow regimes and transitions crucial for advanced hydrological and environmental studies. The model's integration into Landlab's component framework enables future coupling with sediment transport components to simulate morphodynamic processes and assess impacts on aquatic habitat and riverine vegetation dynamics under changing flow conditions.
 
[…]
