### 2025-06-23 124e24d
Update paper.md
Changes to reduce technical description and add an example
