+++ joss/in_preparation/river_flow_dynamics/paper.md
@@ -38,21 +38,9 @@ bibliography: paper.bib
 
 RiverFlowDynamics enables researchers to simulate how water flows through rivers, streams, and flood plains using realistic physics. The software can predict water depths, flow velocities, and how water dynamics evolve over complex terrain, making it valuable for flood risk assessment, environmental studies, and water resource management. Numerical modeling of surface water flow is a critical tool in hydrology, hydraulics, and environmental science. These models play a crucial role in predicting and analyzing flow patterns in rivers, flood plains, and coastal areas, informing decisions in water resource management, flood risk assessment, and ecosystem conservation. This paper introduces a novel two-dimensional flow model, RiverFlowDynamics, developed as a component of the Landlab Python Package [@hobley:2017;@barnhart:2020;@Hutton:2020;@Hutton:2020], designed to simulate the behavior of rivers and streams under various flow conditions over natural and artificial topography.
 
-RiverFlowDynamics is founded on the depth-averaged Saint-Venant equations, also known as the shallow water equations [@casulli1990semi;@casulli_semi-implicit_1999]. These equations, derived from the Navier-Stokes equations for incompressible flow, are simplified by integrating over the water depth. This approach assumes that vertical accelerations are negligible compared to horizontal ones, a reasonable approximation for many surface water flows. The governing equations consist of continuity and momentum balance equations in two dimensions, capturing the essential dynamics of free-surface flows.
+RiverFlowDynamics is founded on the depth-averaged Saint-Venant equations, also known as the shallow water equations [@casulli1990semi;@casulli_semi-implicit_1999], which capture the essential dynamics of free-surface flows by assuming that vertical accelerations are negligible compared to horizontal ones. For the numerical solution, the component employs the finite volume method chosen for its robustness, capacity to handle complex geometries, and inherent conservation properties [@andersson2011computational;@fletcher2012computational], with the model's structure allowing for easy parallelization to enable simulations of extensive river networks or large coastal areas.
 
-For the numerical solution of these equations, RiverFlowDynamics employs the finite volume method, chosen for its robustness, capacity to handle complex geometries, and inherent conservation properties [@andersson2011computational;@fletcher2012computational]. The computational domain is discretized into a uniform rectangular grid, with water surface elevation defined at cell centers and velocity components at cell interfaces. This staggered grid arrangement helps in maintaining numerical stability and accuracy. The numerical implementation results in a penta-diagonal, positive-definite system of equations. This system is solved efficiently using the preconditioned conjugate gradient method, ensuring rapid convergence even for large domains. The model's structure allows for easy parallelization, potentially enabling simulations of extensive river networks or large coastal areas.
-
-A key feature of the model is its semi-implicit and semi-Lagrangian representation. The semi-implicit scheme treats water surface elevation and velocity implicitly, while handling advective terms explicitly. This approach allows for larger time steps compared to fully explicit schemes, enhancing computational efficiency. The semi-Lagrangian method for advection involves tracking fluid particles backwards along their flow lines, providing additional stability and accuracy, particularly for flows with strong advective components [@robert1985semi;@robert_stable_1981]. The model has been validated through a comprehensive test suite including analytical solutions, numerical stability tests, sensitivity analyses, mass conservation checks, and boundary condition validations.
-
[…]
