### 2025-06-23 cf207e1
Adding Figure to Paper
The Figure was added
