### 2025-06-23 d1f39e2
Update paper.md
improving the caption
