### 2025-06-23 997069b
Update paper.md
improving caption 2
