### 2025-06-23 96c4f96
Update paper.md
Reducing to 1000 words from 1246
