### 2025-06-23 7c8ce9e
Update paper.md
getting it to 1000 words part 3
