+++ joss/in_preparation/river_flow_dynamics/paper.md
@@ -36,9 +36,9 @@ bibliography: paper.bib
 ---
 # Summary
 
-RiverFlowDynamics enables researchers to simulate how water flows through rivers, streams, and flood plains using realistic physics, making it valuable for flood risk assessment, environmental studies, and water resource management. This paper introduces the novel two-dimensional flow model developed as a component of the Landlab Python Package [@hobley:2017;@barnhart:2020;@Hutton:2020;@Hutton:2020], designed to simulate river and stream behavior under various flow conditions over natural and artificial topography. RiverFlowDynamics is founded on the depth-averaged Saint-Venant equations, also known as the shallow water equations [@casulli1990semi;@casulli_semi-implicit_1999], which capture essential free-surface flow dynamics by assuming vertical accelerations are negligible compared to horizontal ones. For numerical solution, the component employs the finite volume method chosen for its robustness, capacity to handle complex geometries, and inherent conservation properties [@andersson2011computational;@fletcher2012computational], with the model's structure allowing for easy parallelization to enable simulations of extensive river networks or large coastal areas.
+RiverFlowDynamics enables researchers to simulate how water flows through rivers, streams, and flood plains using realistic physics, making it valuable for flood risk assessment, environmental studies, and water resource management. Numerical modeling of surface water flow is a critical tool in hydrology, hydraulics, and environmental science. These models play a crucial role in predicting and analyzing flow patterns in rivers, flood plains, and coastal areas, informing decisions in water resource management, flood risk assessment, and ecosystem conservation. This paper introduces the novel two-dimensional flow model developed as a component of the Landlab Python Package [@hobley:2017;@barnhart:2020;@Hutton:2020;@Hutton:2020], designed to simulate river and stream behavior under various flow conditions over natural and artificial topography. RiverFlowDynamics is founded on the depth-averaged Saint-Venant equations, also known as the shallow water equations [@casulli1990semi;@casulli_semi-implicit_1999], which capture essential free-surface flow dynamics by assuming vertical accelerations are negligible compared to horizontal ones. For numerical solution, the component employs the finite volume method chosen for its robustness, capacity to handle complex geometries, and inherent conservation properties [@andersson2011computational;@fletcher2012computational], with the model's structure allowing for easy parallelization to enable simulations of extensive river networks or large coastal areas.
 
-A key feature of the model is its semi-implicit and semi-Lagrangian representation that treats water surface elevation and velocity implicitly while handling advective terms explicitly, allowing for larger time steps and enhanced computational efficiency [@robert1985semi;@robert_stable_1981]. The semi-Lagrangian advection scheme tracks fluid particles backwards using Pollock's semi-analytical method [@pollock1988semianalytical], which assumes linear velocity variation within each grid cell and relaxes the Courant-Friedrichs-Lewy condition [@staniforth1991semi;@bates1982multiply]. Source terms account for bottom friction using the Manning-Chezy formula [@he2017numerical;@brufau2000two], while the model features robust dry/wet cell transitions using the method of @casulli1992semi and implements Dirichlet conditions for inlet boundaries and gradient-based or radiation-based conditions for open boundaries. The model has been validated through comprehensive testing including analytical solutions, numerical stability tests, sensitivity analyses, mass conservation checks, and boundary condition validations.
[…]
