### 2025-06-23 ad6f437
Update paper.md
Refining statement of need
