+++ joss/in_preparation/river_flow_dynamics/paper.md
@@ -46,7 +46,7 @@ RiverFlowDynamics offers advantages over simpler overland flow models available 
 
 RiverFlowDynamics is a Python-based 2D flow model developed as a component of the Landlab framework, addressing a critical gap in modeling complex river systems and flood dynamics. Prior to RiverFlowDynamics, Landlab lacked a comprehensive 2D flow model capable of handling fully advective-dominated problems, particularly in rivers with complex topographies, hindering accurate simulations of diverse flow regimes crucial for advanced hydrological and environmental studies. The model's integration into Landlab's component framework enables future coupling with sediment transport components to simulate morphodynamic processes and assess impacts on aquatic habitat and riverine vegetation dynamics under changing flow conditions.
 
-Compared to existing hydraulic modeling software such as TELEMAC and Delft3D, RiverFlowDynamics offers unique advantages through its native Python implementation and integration with the Landlab framework, providing enhanced interoperability with other Earth surface process models and the broader scientific Python ecosystem while maintaining comparable numerical accuracy. The model solves complete depth-averaged Saint-Venant equations, enabling capture of complex flow dynamics including subcritical and supercritical flows, hydraulic jumps, and channel-floodplain interactions, making it valuable for applications from small-scale stream dynamics to large-scale flood simulations, rapid flood propagation in urban areas, detailed floodplain mapping, and complex river morphodynamics analysis in mountainous regions. Integration with Landlab leverages its established grid structure, visualization tools, and interoperability with other components, facilitating multi-process simulations and reducing development overhead.
+Compared to existing hydraulic modeling software such as TELEMAC and Delft3D, RiverFlowDynamics offers unique advantages through its native Python implementation and integration with the Landlab framework, providing enhanced interoperability with other Earth surface process models and the broader scientific Python ecosystem while maintaining comparable numerical accuracy. The model solves complete depth-averaged Saint-Venant equations, enabling capture of complex flow dynamics including subcritical and supercritical flows, hydraulic jumps, and channel-floodplain interactions, making it valuable for applications from small-scale stream dynamics to large-scale flood simulations, rapid flood propagation in urban areas, detailed floodplain mapping, and complex river morphodynamics analysis in mountainous regions. The component's open-source nature and Python-based architecture facilitate educational use and research collaboration, while its integration with Landlab's existing visualization and analysis tools streamlines workflow development. Integration with Landlab leverages its established grid structure, visualization tools, and interoperability with other components, facilitating multi-process simulations and reducing development overhead for researchers studying coupled Earth surface processes.
 
 
 ## Basic Usage Example
