### 2025-06-23 3ec6610
Update paper.md
Submitting for review
