+++ joss/in_preparation/river_flow_dynamics/paper.md
@@ -36,7 +36,7 @@ bibliography: paper.bib
 ---
 # Summary
 
-RiverFlowDynamics enables researchers to simulate how water flows through rivers, streams, and flood plains using realistic physics, making it valuable for flood risk assessment, environmental studies, and water resource management. Numerical modeling of surface water flow is a critical tool in hydrology, hydraulics, and environmental science. These models play a crucial role in predicting and analyzing flow patterns in rivers, flood plains, and coastal areas, informing decisions in water resource management, flood risk assessment, and ecosystem conservation. This paper introduces the novel two-dimensional flow model developed as a component of the Landlab Python Package [@hobley:2017;@barnhart:2020;@Hutton:2020;@Hutton:2020], designed to simulate river and stream behavior under various flow conditions over natural and artificial topography. RiverFlowDynamics is founded on the depth-averaged Saint-Venant equations, also known as the shallow water equations [@casulli1990semi;@casulli_semi-implicit_1999], which capture essential free-surface flow dynamics by assuming vertical accelerations are negligible compared to horizontal ones. For numerical solution, the component employs the finite volume method chosen for its robustness, capacity to handle complex geometries, and inherent conservation properties [@andersson2011computational;@fletcher2012computational], with the model's structure allowing for easy parallelization to enable simulations of extensive river networks or large coastal areas.
+RiverFlowDynamics enables researchers, practitioners, and specialists to simulate how water flows through rivers, streams, and flood plains using realistic physics, making it valuable for flood risk assessment, environmental studies, and water resource management. Numerical modeling of surface water flow is a critical tool in hydrology, hydraulics, and environmental science. These models play a crucial role in predicting and analyzing flow patterns in rivers, flood plains, and coastal areas, informing decisions in water resource management, flood risk assessment, and ecosystem conservation. This paper introduces the novel two-dimensional flow model developed as a component of the Landlab Python Package [@hobley:2017;@barnhart:2020;@Hutton:2020;@Hutton:2020], designed to simulate river and stream behavior under various flow conditions over natural and artificial topography. RiverFlowDynamics is founded on the depth-averaged Saint-Venant equations, also known as the shallow water equations [@casulli1990semi;@casulli_semi-implicit_1999], which capture essential free-surface flow dynamics by assuming vertical accelerations are negligible compared to horizontal ones. For numerical solution, the component employs the finite volume method chosen for its robustness, capacity to handle complex geometries, and inherent conservation properties [@andersson2011computational;@fletcher2012computational], with the model's structure allowing for easy parallelization to enable simulations of extensive river networks or large coastal areas.
 
[…]
