### 2025-06-23 #7823 @angelmons
@cheginit Thank you for your detailed feedback and guidance. I have thoroughly addressed all the points you raised:

**✅ High-level functionality description added:** The paper now begins with a clear, non-specialist description: "RiverFlowDynamics enables researchers, practitioners, and specialists to simulate how water flows through rivers, streams, and flood plains using realistic physics, making it valuable for flood risk assessment, environmental studies, and water resource management."

**✅ Technical details reduced:** I have streamlined and condensed the technical methodology sections while preserving essential information and all references, reducing the overall word count to stay within JOSS's 1000-word limit.

**✅ Brief demonstration section added:** A new "Basic Usage Example" section has been included with:
- A complete, functional code example showing how to set up and run a basic channel flow simulation
- Proper Python syntax highlighting in a code block
- A sample figure (Figure 1) showing the model output with time evolution of water surface elevation

**✅ Sample plot included:** Figure 1 demonstrates the software's capabilities by showing the progression from dry initial conditions to steady-state flow, similar to the example you referenced.

The paper now follows the JOSS format requirements, balances technical accuracy with accessibility for non-specialists, and includes a practical demonstration of the software's functionality. I believe the article is now ready for final review.

Thank you for your patience and constructive feedback throughout this process.
