### 2025-06-24 #7823 @cheginit
@angelmons thanks for addressing the issue. The paper looks good now.

Does v1.0 contain all the fixes that you made to address the comments raised by the reviewers? If not, please release a new version to be associated with your submission for publication. Once you released a new version, provide it to me and generate an archive version of your submission using, e.g., Zenodo, and give me its DOI. For the archive, pleas make sure that the authors, title, and abstract match your submission to JOSS.
