### 2025-06-24 #7823 @cheginit
@angelmons Thanks for confirming the release version. However, the archive that you created does meet the criteria that I asked for in the previous comment:

> For the archive, please make sure that the authors, their ORCiDs, title, and abstract match your submission to JOSS.

Please address this issue with the Zenodo archive. The archive should be specific to this submission, not `landlab`.
