### 2025-06-24 #7823 @angelmons
@cheginit Thank you for the clarification. I have now created a JOSS-specific Zenodo archive that meets your criteria.

**Updated Zenodo Archive:**

The new archive is specifically for this JOSS submission with metadata that matches our paper exactly:

- **Repository**: angelmons/RiverFlowDynamics-JOSS
- **Release**: V1.0
- **DOI**: https://doi.org/10.5281/zenodo.15731998

**Verification of matching metadata:**
- ✅ **Title**: "RiverFlowDynamics v1.0: A Landlab component for computing two-dimensional river flow dynamics"
- ✅ **Authors**: Angel Monsalve, Sebastian Bernal, Oscar Link
- ✅ **ORCIDs**: All author ORCIDs match the JOSS submission
- ✅ **Abstract**: Consistent with the JOSS paper summary

The archive is now specific to this submission rather than the entire Landlab repository, and all metadata aligns perfectly with our JOSS submission.

Thank you for your patience in helping us get this right!
