### 2025-06-24 #7823 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
