### 2025-06-24 #7823 @kthyng
Congratulations on your new publication @angelmons! Many thanks to editor @cheginit and to reviewers @mcflugen, @ZhiLiHydro and @xiaxilin for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
