### 2025-06-24 #7823 @angelmons
Thank you so much @kthyng

I just signed up as a reviewer. I will be a pleasure to contribute to JOSS
